backend/app/requirements.txt
fastapi>=0.100.0
uvicorn==0.22.0
langchain==0.1.0
pydantic>=2.7.0
openai==0.27.7
python-dotenv==1.0.0
black==23.3.0
chromadb==0.4.24
langchain-community==0.0.20
langchain-core>=0.1.7,<0.2.0
sentence-transformers>=2.2.2
